requirements.txt
numpy
scipy
Pillow
matplotlib
scikit-learn
tqdm
datasets
huggingface_hub
